Move subscription logic from provider to lcm module.
This is logic that is not specific to any one provider. This also
eliminates the lifetimes that had to be specified by the provider struct.
